refactor(server): Moved MediaTypesRule logic out of PatternUtil #950
