+++ server/src/main/java/de/zalando/zally/util/PatternUtil.kt
@@ -6,16 +6,8 @@ package de.zalando.zally.util
 object PatternUtil {
 
     private val PATH_VARIABLE_PATTERN = "\\{.+}$".toRegex()
-    private val APPLICATION_PROBLEM_JSON_PATTERN = "^application/(problem\\+)?json$".toRegex()
-    private val CUSTOM_WITH_VERSIONING_PATTERN = "^\\w+/[-+.\\w]+;v(ersion)?=\\d+$".toRegex()
 
     fun hasTrailingSlash(input: String): Boolean = input.trim { it <= ' ' }.endsWith("/")
 
     fun isPathVariable(input: String): Boolean = input.matches(PATH_VARIABLE_PATTERN)
-
-    fun isApplicationJsonOrProblemJson(mediaType: String): Boolean =
-        mediaType.matches(APPLICATION_PROBLEM_JSON_PATTERN)
-
-    fun isCustomMediaTypeWithVersioning(mediaType: String): Boolean =
-        mediaType.matches(CUSTOM_WITH_VERSIONING_PATTERN)
 }
